Refactor: 입력과 관련된 테스트 코드 분리
